+++ src/particles/deposition/PlasmaDepositCurrent.cpp
@@ -7,6 +7,7 @@
  */
 #include "PlasmaDepositCurrent.H"
 
+#include "DepositionUtil.H"
 #include "particles/particles_utils/ShapeFactors.H"
 #include "particles/particles_utils/FieldGather.H"
 #include "particles/plasma/PlasmaParticleContainer.H"
@@ -63,9 +64,6 @@ DepositCurrent (PlasmaParticleContainer& plasma, Fields & fields,
         const amrex::Real x_pos_offset = GetPosOffset(0, gm[lev], isl_fab.box());
         const amrex::Real y_pos_offset = GetPosOffset(1, gm[lev], isl_fab.box());
 
-        // Extract particle properties
-        const auto ptd = pti.GetParticleTile().getParticleTileData();
-
         // Extract box properties
         const amrex::Real dx_inv = gm[lev].InvCellSize(0);
         const amrex::Real dy_inv = gm[lev].InvCellSize(1);
@@ -91,161 +89,155 @@ DepositCurrent (PlasmaParticleContainer& plasma, Fields & fields,
         AMREX_ALWAYS_ASSERT_WITH_MESSAGE(isl_fab.box().ixType().cellCentered(),
             "jx, jy, jz, and rho must be cell centered in all directions.");
 
-        Array3<amrex::Real> const isl_arr = isl_fab.array();
-
-        const bool do_tiling = Hipace::m_do_tiling;
-
-        [[maybe_unused]] int ntilex = 1;
-        [[maybe_unused]] int ntiley = 1;
-        if (do_tiling) {
-            AMREX_ALWAYS_ASSERT_WITH_MESSAGE(bin_size >= 2*Fields::m_slices_nguards[0],
-            "plasmas.sort_bin_size must be at least twice as large as the number of ghost cells");
-            amrex::Box domain = gm[lev].Domain();
-            domain.coarsen(bin_size);
-            ntilex = domain.length(0);
-            ntiley = domain.length(1);
-        }
-
-#ifndef AMREX_USE_GPU
-        for (int tile_perm_x=0; tile_perm_x<2; ++tile_perm_x) {
-        for (int tile_perm_y=0; tile_perm_y<2; ++tile_perm_y) {
-#ifdef AMREX_USE_OMP
-#pragma omp parallel for collapse(2) if(do_tiling)
-#endif
-        for (int itilex=tile_perm_x; itilex<ntilex; itilex+=2) {
-        for (int itiley=tile_perm_y; itiley<ntiley; itiley+=2) {
-        // the index is transposed to be the same as in amrex::DenseBins::build
-        const int tile_index = itilex * ntiley + itiley;
-        using tiling_cto = amrex::CompileTimeOptions<false, true>;
-#else
-        const int tile_index = 0;
-        using tiling_cto = amrex::CompileTimeOptions<false>; // tiling disabled on GPU
-#endif
-
-        PlasmaBins::index_type const * const a_indices = do_tiling ? bins.permutationPtr() : nullptr;
-        PlasmaBins::index_type const * const a_offsets = do_tiling ? bins.offsetsPtr() : nullptr;
-
-        int num_particles = do_tiling ? a_offsets[tile_index+1]-a_offsets[tile_index]
-                                        : pti.numParticles();
+        Array3<amrex::Real> const field_arr = isl_fab.array();
 
         // Loop over particles and deposit into jx_fab, jy_fab, jz_fab, and rho_fab
-        amrex::ParallelFor(
+        amrex::AnyCTO(
             amrex::TypeList<
                 amrex::CompileTimeOptions<0, 1, 2, 3>,  // depos_order
                 amrex::CompileTimeOptions<false, true>, // can_ionize
-                tiling_cto,                             // do_tiling
                 amrex::CompileTimeOptions<false, true>  // use_laser
             >{}, {
                 Hipace::m_depos_order_xy,
                 plasma.m_can_ionize,
-                do_tiling,
                 Hipace::m_use_laser
             },
-            num_particles,
-            [=] AMREX_GPU_DEVICE (int ip, auto depos_order,
-                auto can_ionize, auto c_do_tiling, auto use_laser) noexcept {
-            constexpr int depos_order_xy = depos_order.value;
-
-            [[maybe_unused]] auto indices = a_indices;
-            [[maybe_unused]] auto offsets = a_offsets;
-            [[maybe_unused]] auto itile = tile_index;
-            if constexpr (c_do_tiling.value) {
-                ip = indices[offsets[itile]+ip];
-            }
-
-            // only deposit plasma currents on or below their according MR level
-            if (!ptd.id(ip).is_valid() || (lev != 0 && ptd.cpu(ip) < lev)) return;
-
-            const amrex::Real psi_inv = 1._rt/ptd.rdata(PlasmaIdx::psi)[ip];
-            const amrex::Real xp = ptd.pos(0, ip);
-            const amrex::Real yp = ptd.pos(1, ip);
-            const amrex::Real vx_c = ptd.rdata(PlasmaIdx::ux)[ip] * psi_inv;
-            const amrex::Real vy_c = ptd.rdata(PlasmaIdx::uy)[ip] * psi_inv;
-
-            // calculate charge of the plasma particles
-            amrex::Real q_invvol = charge_invvol * ptd.rdata(PlasmaIdx::w)[ip];
-            amrex::Real q_mu0_mass_ratio = charge_mu0_mass_ratio;
-            [[maybe_unused]] amrex::Real laser_norm_ion = laser_norm;
-            if constexpr (can_ionize.value) {
-                q_invvol *= ptd.idata(PlasmaIdx::ion_lev)[ip];
-                q_mu0_mass_ratio *= ptd.idata(PlasmaIdx::ion_lev)[ip];
-                laser_norm_ion *=
-                    ptd.idata(PlasmaIdx::ion_lev)[ip] * ptd.idata(PlasmaIdx::ion_lev)[ip];
-            }
-
-            const amrex::Real xmid = (xp - x_pos_offset) * dx_inv;
-            const amrex::Real ymid = (yp - y_pos_offset) * dy_inv;
-
-            amrex::Real Aabssqp = 0._rt;
-            [[maybe_unused]] auto a_isl_arr = isl_arr;
-            [[maybe_unused]] auto a_aabs = aabs;
-            if constexpr (use_laser.value) {
-                doLaserGatherShapeN<depos_order_xy>(xp, yp, Aabssqp, a_isl_arr, a_aabs,
-                                                    dx_inv, dy_inv, x_pos_offset, y_pos_offset);
-                Aabssqp *= laser_norm_ion;
-            }
-
-            // calculate gamma/psi for plasma particles
-            const amrex::Real gamma_psi = 0.5_rt * (
-                (1._rt + 0.5_rt * Aabssqp) * psi_inv * psi_inv
-                + vx_c * vx_c * clightinv * clightinv
-                + vy_c * vy_c * clightinv * clightinv
-                + 1._rt
-            );
-
-            if (gamma_psi < 0.0_rt || gamma_psi > max_qsa_weighting_factor || psi_inv < 0.0_rt)
+            [&](auto is_valid, auto get_cell, auto deposit){
+                constexpr auto ctos = deposit.GetOptions();
+                constexpr int depos_order = ctos[0];
+                constexpr int use_laser = ctos[2];
+                constexpr int stencil_size = depos_order + 1;
+                if constexpr (use_laser) {
+                    SharedMemoryDeposition<stencil_size, stencil_size>(
+                        int(pti.numParticles()), is_valid, get_cell, deposit, field_arr,
+                        isl_fab.box(), pti.GetParticleTile().getParticleTileData(),
+                        std::array{aabs}, std::array{jx, jy, jz, rho, chi, rhomjz});
+                } else {
+                    SharedMemoryDeposition<stencil_size, stencil_size>(
+                        int(pti.numParticles()), is_valid, get_cell, deposit, field_arr,
+                        isl_fab.box(), pti.GetParticleTile().getParticleTileData(),
+                        std::array<int, 0>{}, std::array{jx, jy, jz, rho, chi, rhomjz});
+                }
+            },
+            [=] AMREX_GPU_DEVICE (int ip, auto ptd,
+                                  auto depos_order, auto can_ionize,
+                                  auto use_laser)
             {
-                // This particle violates the QSA, discard it and do not deposit its current
-                amrex::Gpu::Atomic::Add(p_n_qsa_violation, 1);
-                ptd.rdata(PlasmaIdx::w)[ip] = 0.0_rt;
-                ptd.id(ip).make_invalid();
-                return;
-            }
+                // only deposit plasma currents on or below their according MR level
+                return ptd.id(ip).is_valid() && (lev == 0 || ptd.cpu(ip) >= lev);
+            },
+            [=] AMREX_GPU_DEVICE (int ip, auto ptd,
+                                  auto depos_order, auto can_ionize,
+                                  auto use_laser) -> amrex::IntVectND<2>
+            {
+                const amrex::Real xp = ptd.pos(0, ip);
+                const amrex::Real yp = ptd.pos(1, ip);
 
-            for (int iy=0; iy <= depos_order_xy; ++iy) {
-                for (int ix=0; ix <= depos_order_xy; ++ix) {
+                const amrex::Real xmid = (xp - x_pos_offset) * dx_inv;
+                const amrex::Real ymid = (yp - y_pos_offset) * dy_inv;
 
-                    // --- Compute shape factors
-                    // x direction
-                    auto [shape_x, cell_x] =
-                        compute_single_shape_factor<false, depos_order_xy>(xmid, ix);
+                auto [shape_x, i] =
+                    compute_single_shape_factor<false, depos_order>(xmid, 0);
 
-                    // y direction
-                    auto [shape_y, cell_y] =
-                        compute_single_shape_factor<false, depos_order_xy>(ymid, iy);
+                auto [shape_y, j] =
+                    compute_single_shape_factor<false, depos_order>(ymid, 0);
 
-                    const amrex::Real charge_density = q_invvol * shape_x * shape_y;
-                    // wqx, wqy wqz are particle current in each direction
-                    const amrex::Real wqx     = charge_density * vx_c;
-                    const amrex::Real wqy     = charge_density * vy_c;
-                    const amrex::Real wqz     = charge_density * (gamma_psi-1._rt) * clight;
-                    const amrex::Real wq      = charge_density * gamma_psi;
-                    const amrex::Real wchi    = charge_density * q_mu0_mass_ratio * psi_inv;
-                    const amrex::Real wrhomjz = charge_density;
+                return {i, j};
+            },
+            [=] AMREX_GPU_DEVICE (int ip, auto ptd,
+                                  Array3<amrex::Real> arr,
+                                  auto cache_idx, auto depos_idx,
+                                  auto depos_order, auto can_ionize,
+                                  auto use_laser) noexcept
+            {
+                // only deposit plasma currents on or below their according MR level
+                if (!ptd.id(ip).is_valid() || (lev != 0 && ptd.cpu(ip) < lev)) return;
 
-                    // Deposit current into isl_arr
-                    if (jx != -1) { // deposit_jx_jy
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jx), wqx);
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jy), wqy);
-                    }
-                    if (jz != -1) { // deposit_jz
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, jz), wqz);
-                    }
-                    if (rho != -1) { // deposit_rho
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, rho), wq);
-                    }
-                    if (chi != -1) { // deposit_chi
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, chi), wchi);
-                    }
-                    if (rhomjz != -1) { // deposit_rhomjz
-                        amrex::Gpu::Atomic::Add(isl_arr.ptr(cell_x, cell_y, rhomjz), wrhomjz);
+                const amrex::Real psi_inv = 1._rt/ptd.rdata(PlasmaIdx::psi)[ip];
+                const amrex::Real xp = ptd.pos(0, ip);
+                const amrex::Real yp = ptd.pos(1, ip);
+                const amrex::Real vx_c = ptd.rdata(PlasmaIdx::ux)[ip] * psi_inv;
+                const amrex::Real vy_c = ptd.rdata(PlasmaIdx::uy)[ip] * psi_inv;
+
+                // calculate charge of the plasma particles
+                amrex::Real q_invvol = charge_invvol * ptd.rdata(PlasmaIdx::w)[ip];
+                amrex::Real q_mu0_mass_ratio = charge_mu0_mass_ratio;
+                [[maybe_unused]] amrex::Real laser_norm_ion = laser_norm;
+                if constexpr (can_ionize) {
+                    q_invvol *= ptd.idata(PlasmaIdx::ion_lev)[ip];
+                    q_mu0_mass_ratio *= ptd.idata(PlasmaIdx::ion_lev)[ip];
+                    laser_norm_ion *=
+                        ptd.idata(PlasmaIdx::ion_lev)[ip] * ptd.idata(PlasmaIdx::ion_lev)[ip];
+                }
+
+                const amrex::Real xmid = (xp - x_pos_offset) * dx_inv;
+                const amrex::Real ymid = (yp - y_pos_offset) * dy_inv;
+
+                amrex::Real Aabssqp = 0._rt;
+                if constexpr (use_laser) {
+                    doLaserGatherShapeN<depos_order>(xp, yp, Aabssqp, arr, cache_idx[0],
+                                                    dx_inv, dy_inv, x_pos_offset, y_pos_offset);
+                    Aabssqp *= laser_norm_ion;
+                }
+
+                // calculate gamma/psi for plasma particles
+                const amrex::Real gamma_psi = 0.5_rt * (
+                    (1._rt + 0.5_rt * Aabssqp) * psi_inv * psi_inv
+                    + vx_c * vx_c * clightinv * clightinv
+                    + vy_c * vy_c * clightinv * clightinv
+                    + 1._rt
+                );
+
+                if (gamma_psi < 0.0_rt || gamma_psi > max_qsa_weighting_factor || psi_inv < 0.0_rt)
+                {
+                    // This particle violates the QSA, discard it and do not deposit its current
+                    amrex::Gpu::Atomic::Add(p_n_qsa_violation, 1);
+                    ptd.rdata(PlasmaIdx::w)[ip] = 0.0_rt;
+                    ptd.id(ip).make_invalid();
+                    return;
+                }
+
+                for (int iy=0; iy <= depos_order; ++iy) {
+                    for (int ix=0; ix <= depos_order; ++ix) {
+
+                        // --- Compute shape factors
+                        // x direction
+                        auto [shape_x, i] =
+                            compute_single_shape_factor<false, depos_order>(xmid, ix);
+
+                        // y direction
+                        auto [shape_y, j] =
+                            compute_single_shape_factor<false, depos_order>(ymid, iy);
+
+                        const amrex::Real charge_density = q_invvol * shape_x * shape_y;
+                        // wqx, wqy wqz are particle current in each direction
+                        const amrex::Real wqx     = charge_density * vx_c;
+                        const amrex::Real wqy     = charge_density * vy_c;
+                        const amrex::Real wqz     = charge_density * (gamma_psi-1._rt) * clight;
+                        const amrex::Real wq      = charge_density * gamma_psi;
+                        const amrex::Real wchi    = charge_density * q_mu0_mass_ratio * psi_inv;
+                        const amrex::Real wrhomjz = charge_density;
+
+                        // Deposit current into arr
+                        if (depos_idx[0] != -1) { // deposit_jx_jy
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[0]), wqx);
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[1]), wqy);
+                        }
+                        if (depos_idx[2] != -1) { // deposit_jz
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[2]), wqz);
+                        }
+                        if (depos_idx[3] != -1) { // deposit_rho
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[3]), wq);
+                        }
+                        if (depos_idx[4] != -1) { // deposit_chi
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[4]), wchi);
+                        }
+                        if (depos_idx[5] != -1) { // deposit_rhomjz
+                            amrex::Gpu::Atomic::Add(arr.ptr(i, j, depos_idx[5]), wrhomjz);
+                        }
                     }
                 }
-            }
-        });
-#ifndef AMREX_USE_GPU
-        }}}}
-#endif
+            });
 
         n_qsa_violation = gpu_n_qsa_violation.dataValue();
         if (n_qsa_violation > 0 && (Hipace::m_verbose >= 3))
